fix: 移除消息编码中OperationType的冗余类型转换
移除对OperationType属性的显式byte类型转换，因为该属性已为byte类型，避免不必要的装箱操作。

+++ GameFrameX.NetWork.Message/DefaultMessageEncoderHandler.cs
@@ -71,7 +71,7 @@ public sealed class DefaultMessageEncoderHandler : BaseMessageEncoderHandler
                 // 总长度
                 buffer.WriteUIntValue((uint)totalLength, ref offset);
                 // operationType 操作类型
-                buffer.WriteByteValue((byte)messageObject.OperationType, ref offset);
+                buffer.WriteByteValue(messageObject.OperationType, ref offset);
                 // zipFlag 压缩标记
                 buffer.WriteByteValue(zipFlag, ref offset);
                 // uniqueId 唯一ID
